wygląda na to, że to finał - pull z gałęzi w, której były dodawane wszystkie ostatnie poprawy logiki i wymagań technicznych symulacji, symulacja sprawna w normalnym działaniu, testach ogólnych i testach swoich

+++ src/main.cpp
@@ -10,6 +10,7 @@
 #include <mutex>
 #include <atomic>
 #include <algorithm>
+#include <cerrno>
 #include "operacje.h"
 #include "struktury.h"
 #include "logger.h"
@@ -20,42 +21,60 @@ std::mutex mutex_procesy;
 pid_t kasjer_pid = 0;
 pid_t pracownik_pid = 0;
 volatile int koniec = 0; //flaga do handera
-std::atomic<int> liczba_aktywnych_klientow{0};
+volatile sig_atomic_t przerwanie_uzytkownika = 0;
+std::atomic<int> liczba_aktywnych_klientow{0}; //liczba wygenerowanych procesów klienta
+
+//flagi do poprawnej ewakuacji w logach
+volatile bool kasjer_zakonczony = false;
+volatile bool pracownik_zakonczony = false;
 
 //usuwanie klientów zombie
 void watekSprzatajacy() {
-    while(true) {
-        if (koniec) {
-            return;
-        }
-        
-        int status;
-        while (true) {
-            pid_t zakonczony = waitpid(-1, &status, WNOHANG);
-            if (zakonczony > 0) {
-                std::lock_guard<std::mutex> lock(mutex_procesy);
-                auto it = std::remove(procesy_potomne.begin(), procesy_potomne.end(), zakonczony);
-                if (it != procesy_potomne.end()) {
-                    procesy_potomne.erase(it, procesy_potomne.end());
-                    liczba_aktywnych_klientow--;
-                }
+    int status;
+    while (true) {
+        pid_t zakonczony = waitpid(-1, &status, 0);
+        if (zakonczony > 0) {
+            std::lock_guard<std::mutex> lock(mutex_procesy);
+
+            if (zakonczony == kasjer_pid) {
+                kasjer_zakonczony = true;
+            }
+            else if (zakonczony == pracownik_pid) {
+                pracownik_zakonczony = true;
             }
             else {
-                break;
+                for (size_t i = 0; i < procesy_potomne.size(); ++i) {
+                    if (procesy_potomne[i] == zakonczony) {
+                        procesy_potomne[i] = procesy_potomne.back();
+                        procesy_potomne.pop_back();
+                        liczba_aktywnych_klientow--;
+                        break;
+                    }
+                }
             }
+            continue;
         }
-        usleep(50000);
+
+        if (zakonczony == -1) {
+            if (errno == EINTR) {
+                continue;
+            }
+            break;
+        }
     }
 }
 
 //handler sygnału przerwania symulacji
 void przerwanie(int sig) {
     koniec = 1;
+    przerwanie_uzytkownika = 1;
 }
 
 int main() {
     signal(SIGINT, przerwanie);
-    signal(SIGTERM, przerwanie);
+    signal(SIGTERM, SIG_IGN);
+    signal(SIGUSR1, SIG_IGN);
+    signal(SIGUSR2, SIG_IGN);
     srand(time(NULL));
 
     tabula_rasa();
@@ -69,28 +88,28 @@ int main() {
     //inicjalizacja elementów z pamięci dzielonej
     semaforOpusc(sem_id, SEM_STOLIKI);
     for (int i = 0; i < STOLIKI_X1; i++) {
-        pam->stoliki_x1[i].id = i + 1;
+        pam->stoliki_x1[i].id = i;
         pam->stoliki_x1[i].pojemnosc_max = 1;
         pam->stoliki_x1[i].ile_zajetych_miejsc = 0;
         pam->stoliki_x1[i].wielkosc_grupy_siedzacej = 0;
         pam->stoliki_x1[i].zarezerwowany = false;
     }
     for (int i = 0; i < STOLIKI_X2; i++) {
-        pam->stoliki_x2[i].id = i + 1;
+        pam->stoliki_x2[i].id = i;
         pam->stoliki_x2[i].pojemnosc_max = 2;
         pam->stoliki_x2[i].ile_zajetych_miejsc = 0;
         pam->stoliki_x2[i].wielkosc_grupy_siedzacej = 0;
         pam->stoliki_x2[i].zarezerwowany = false;
     }
     for (int i = 0; i < STOLIKI_X3; i++) {
-        pam->stoliki_x3[i].id = i + 1;
+        pam->stoliki_x3[i].id = i;
         pam->stoliki_x3[i].pojemnosc_max = 3;
         pam->stoliki_x3[i].ile_zajetych_miejsc = 0;
         pam->stoliki_x3[i].wielkosc_grupy_siedzacej = 0;
         pam->stoliki_x3[i].zarezerwowany = false;
     }
     for (int i = 0; i < STOLIKI_X4; i++) {
-        pam->stoliki_x4[i].id = i + 1;
+        pam->stoliki_x4[i].id = i;
         pam->stoliki_x4[i].pojemnosc_max = 4;
         pam->stoliki_x4[i].ile_zajetych_miejsc = 0;
         pam->stoliki_x4[i].wielkosc_grupy_siedzacej = 0;
@@ -105,10 +124,11 @@ int main() {
     pam->aktualna_liczba_X3 = STOLIKI_X3 / 2;
     pam->liczba_klientow = 0;
     pam->utarg = 0;
+    pam->pracownik_pid = 0;
+    pam->pgid_grupy = getpgrp();
     semaforPodnies(sem_id, SEM_MAIN);
 
-    std::string zasoby = "Utworzono zasoby";
-    logger(zasoby);
+    logger("Utworzono zasoby, start symulacji pgid: " + std::to_string(pam->pgid_grupy));
     
     pid_t pid;
 
@@ -140,13 +160,13 @@ int main() {
 
     //kierownik odpalany manualnie w osobnej konsoli do wydawania poleceń
 
-    usleep(1000000); //opóźnienie dla estetyki w konsoli, żeby kasjer i pracownik byli pierwsi
+    usleep(500000); //opóźnienie dla estetyki w konsoli, żeby kasjer i pracownik byli pierwsi
     //klienci
     while(!koniec) {
-        
         semaforOpusc(sem_id, SEM_MAIN);
         if (pam->pozar) {
             semaforPodnies(sem_id, SEM_MAIN);
+            koniec = 1;
             break;
         }
         else {
@@ -167,71 +187,51 @@ int main() {
                 std::lock_guard<std::mutex> lock(mutex_procesy);
                 procesy_potomne.push_back(pid);
                 liczba_aktywnych_klientow++;
-                usleep(1000);
-            }
-            else {
-                if (errno == EAGAIN) {
-                    usleep(50000);
-                    continue;
-                }
-                else {
-                    perror("Błąd fork");
-                    break;
-                }
             }
         }
         else {
-            usleep(10000);
+            usleep(500000 + (rand() % 2000000)); //losowy czas pojawiania się klientów
         }
     }
 
-    //oczekiwanie na sygnały
-    while(true) {
-        if (koniec) {
-            break;
-        }
-        semaforOpusc(sem_id, SEM_MAIN);
-        if (pam->pozar) {
-            semaforPodnies(sem_id, SEM_MAIN);
-            int status_kasjer = kill(kasjer_pid, 0);
-            int status_pracownik = kill(pracownik_pid, 0);
-            if (status_kasjer == -1 && errno == ESRCH &&
-                status_pracownik == -1 && errno == ESRCH) {
-                    usleep(500000);
-                    break;
-            }
-        }
-        else {
-            semaforPodnies(sem_id, SEM_MAIN);
-        }
-        usleep(1000000);
+    if (przerwanie_uzytkownika) {
+        logger("Symulacja: Koniec tworzenia klientów");
+    }
+    else {
+        logger("Symulacja: Koniec tworzenia klientów. Oczekiwanie na koniec ewakuacji");
     }
 
-    semaforOpusc(sem_id, SEM_MAIN);
-    long ostateczny_utarg = pam->utarg;
-    semaforPodnies(sem_id, SEM_MAIN);
-
-    //sprzątanie po symulacji
-    {
+    //zabezpieczenie dobicia procesów
+    if (przerwanie_uzytkownika) {
+        if (kasjer_pid > 0) kill(kasjer_pid, SIGKILL);
+        if (pracownik_pid > 0) kill(pracownik_pid, SIGKILL);
         std::lock_guard<std::mutex> lock(mutex_procesy);
         for (pid_t pid : procesy_potomne) {
             kill(pid, SIGKILL);
         }
     }
+    while (true) {
+        int status;
+        pid_t z = waitpid(-1, &status, 0);
+        if (z > 0) {
+            continue;
+        }
+        if (z == -1 && errno == EINTR) {
+            continue;
+        }
+        break;
+    }
 
-    if (kasjer_pid > 0) {
-        kill(kasjer_pid, SIGKILL);
-    }
-    if (pracownik_pid > 0) {
-        kill(pracownik_pid, SIGKILL);
-    }
+    //sprzątnięcie IPC i podsumowanie symulacji
+    semaforOpusc(sem_id, SEM_MAIN);
+    long ostateczny_utarg = pam->utarg;
+    semaforPodnies(sem_id, SEM_MAIN);
 
     usunSemafor(sem_id);
     usunKolejke(kol_id);
     zwolnijPamiec(pam_id);
 
-    std::string utarg_info = "Koniec dnia. Utarg wyniósł: " + std::to_string(ostateczny_utarg) + " PLN";
-    logger(utarg_info);
+    logger("Koniec dnia. Utarg wyniósł: " + std::to_string(ostateczny_utarg) + " PLN");
     logger("Symulacja zakończona i wszystko sprzątnięte");
     return 0;
 }
